Sarsa
Here is what I have for sarsa.  It needs testing if anyone can take a
look at it.  When I ran it nothing seemed to happen.  I'm not sure if it
just takes a long time to learn or if something is wrong.  It could be
slow, there are alot of action combinations it needs to see.  I will
test it more tomorrow.